Panel design (#152)
* Small changes to models
Change TimeZone
Translations and migrations

* Create new panel

* Separated templates

* Separated views and updated urls

* Bug Fix

* Few changes to tests

* conflicts with dev

* translation
battle history pagination
small changes to views

* translation
battle history pagination
small changes to views
