+++ aic_site/settings/base.py
@@ -124,7 +124,7 @@ LANGUAGES = (
 )
 LANGUAGE_CODE = 'fa'
 
-TIME_ZONE = 'UTC'
+TIME_ZONE = 'Asia/Tehran'
 
 USE_I18N = True
