--- apps/accounts/templates/accounts/friendly_match_scoreboard.html
@@ -1,25 +1,0 @@
-{% extends 'base/base.html' %}
-{% load static i18n %}
-{% block container %}
-<h2>{% trans 'Scoreboard' %}</h2>
-    <table class="ui center aligned table table-bordered">
-        <thead>
-            <tr>
-                <th>{% trans 'Rank' %}</th>
-                <th>{% trans 'Name' %}</th>
-                <th>{% trans 'Score' %}</th>
-                <th>{% trans 'total num' %}</th>
-                <th>{% trans 'win num' %}</th>
-                <th>{% trans 'lose num' %}</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for competitor in league_scoreboard %}
-                {% for key, value in competitor.items %}
-                    <tr>{{ value }}</tr>
-                {% endfor %}
-            {% endfor %}
-        </tbody>
-
-    </table>
-{% endblock %}
