+++ apps/accounts/templates/accounts/oldpanel/friendly_result.html
@@ -1,0 +1,21 @@
+{% extends 'base/base.html' %}
+{% load i18n %}
+
+{% block container %}
+    {% if form.errors|length %}
+        <div class="ui error message">
+            <strong>{% trans 'Error scheduling match.' %}</strong>
+            {% for error in form.non_field_errors %}
+                <p>
+                    <br>
+                    <strong>{{ error|escape }}</strong>
+                </p>
+            {% endfor %}
+        </div>
+    {% else %}
+        <div class="ui success message">
+            {% trans 'Successfully scheduled friendly match.' %}
+        </div>
+    {% endif %}
+    <a class="ui primary button" href="{% url 'accounts:panel' %}">{% trans 'Panel' %}</a>
+{% endblock %}
