+++ apps/accounts/templates/accounts/oldpanel/submissions.html
@@ -1,0 +1,152 @@
+{% load i18n %}
+{% load static %}
+{% get_current_language_bidi as LANGUAGE_BIDI %}
+
+
+{% if participation.challenge.is_submission_open or submissions|length > 0 %}
+    <h2>{% trans 'Submissions' %}</h2>
+{% endif %}
+{% if submissions|length > 0 %}
+    <table class="ui center aligned table table-bordered">
+        <thead>
+        <tr>
+            <th>{% trans "Selected" %}</th>
+            <th>{% trans "Time" %}</th>
+            <th>{% trans "Language" %}</th>
+            <th>{% trans "Status" %}</th>
+        </tr>
+        </thead>
+        <tbody>
+        {% for submission in submissions %}
+            <tr>
+                <td>
+                    <div class="ui checkbox">
+                        <input type="checkbox" class="submission"
+                               {% if not submission.team.challenge.is_submission_open or submission.status != 'compiled' %}disabled{% endif %}
+                               {% if submission.is_final %}checked="true"{% endif %}
+                               data-id="{{ submission.id }}" id="submission_{{ submission.id }}">
+                        <label></label>
+                    </div>
+                </td>
+                <style>
+                    {#                                        add_submission_api({{ submission.id }})#}
+                </style>
+                <td><a href="{{ submission.file.url }}">{{ submission.time }}</a></td>
+                <td>
+                    {% if submission.language == 'java' %}
+                        <div class="ui orange label">Java</div>
+                    {% elif submission.language == 'py3' %}
+                        <div class="ui blue label">Python</div>
+                    {% elif submission.language == 'cpp' %}
+                        <div dir="ltr" class="ui purple label">C++</div>
+                    {% endif %}
+                </td>
+                <td>
+                    {% if submission.status == 'compiled' %}
+                        <a class="ui green label" data-tooltip="{% trans 'Compiled Successfully' %}"
+                           data-inverted="">{% trans 'Compiled' %}</a>
+                    {% elif submission.status == 'failed' %}
+                        <a onclick="$('#submission_modal_{{ submission.id }}').show()" class="ui red label"
+                           data-tooltip="{% trans 'Click To View' %}" data-inverted="">{% trans 'Failed' %}</a>
+                        <div class="ui modal" id="submission_modal_{{ submission.id }}">
+                            <div class="header">
+                                {% trans 'Compile Error' %}
+                            </div>
+                            <i class="close icon"></i>
+                            <div class="scrolling content">
+                                <div class="description">
+                                    <div class="ui header">
+                                        {% trans 'The Log From Compiler Is Shown Below:' %}
+                                    </div>
+                                    <p dir="ltr" style="float: left;">
+                                        {{ submission.infra_compile_message }}
+                                    </p>
+                                </div>
+                            </div>
+                            <div class="actions">
+                                <div class="ui black deny button"
+                                     onclick="$('#submission_modal_{{ submission.id }}').hide()">
+                                    {% trans 'OK' %}
+                                </div>
+                            </div>
+                        </div>
+                    {% elif submission.status == 'uploading' %}
+                        <a class="ui label"
+                           data-tooltip="{% trans 'Failed to upload. You may contact the technical team.' %}"
+                           data-inverted="">{% trans 'Failed To Upload' %}</a>
+                    {% elif submission.status == 'compiling' %}
+                        <a class="ui blue label"
+                           data-tooltip="{% trans 'We are compiling your code! Wait and refresh your page.' %}"
+                           data-inverted="">{% trans 'Compiling' %}</a>
+                    {% endif %}
+                </td>
+            </tr>
+        {% endfor %}
+        </tbody>
+        {% if submissions.has_other_pages %}
+            <tfoot>
+            <tr>
+                <th colspan="4">
+                    <div class="ui right floated pagination menu">
+                        {% if submissions.has_previous %}
+                            <a class="icon item" href="?battles_page={{ battles_page }}&page={{ submissions.previous_page_number }}">
+                                <i class="{% if LANGUAGE_BIDI %}right{% else %}left{% endif %} chevron icon"></i>
+                            </a>
+                        {% else %}
+                            <a class="disabled icon item">
+                                <i class="{% if LANGUAGE_BIDI %}right{% else %}left{% endif %} chevron icon"></i>
+                            </a>
+                        {% endif %}
+                        {% for i in submissions.paginator.page_range %}
+                            {% if submissions.number == i %}
+                                <a class="active item"><span>{{ i }}</span></a>
+                            {% else %}
+                                <a class="item" href="?battles_page={{ battles_page }}&page={{ i }}"><span>{{ i }}</span></a>
+                            {% endif %}
+                        {% endfor %}
+                        {% if submissions.has_next %}
+                            <a class="icon item" href="?battles_page={{ battles_page }}&page={{ submissions.next_page_number }}">
+                                <i class="{% if LANGUAGE_BIDI %}left{% else %}right{% endif %} chevron icon"></i>
+                            </a>
+                        {% else %}
+                            <a class="disabled icon item">
+                                <i class="{% if LANGUAGE_BIDI %}left{% else %}right{% endif %} chevron icon"></i>
+                            </a>
+                        {% endif %}
+                    </div>
+                </th>
+            </tr>
+            </tfoot>
+        {% endif %}
+    </table>
+    {{ form.non_field_errors }}
+{% endif %}
+{% if participation.challenge.is_submission_open %}
+    {% csrf_token %}
+    <div class="one field">
+        <label for="{{ form.file.id_for_label }}">{% trans 'File:' %}</label>
+
+        {{ form.file }}
+
+        {% if form.file.errors %}
+            <div class="ui red message">{{ form.file.errors }}</div>
+        {% endif %}
+    </div>
+    <div class="one field" style="height: 100%;">
+        <label for="{{ form.language.id_for_label }}">{% trans 'Language:' %}</label>
+
+        <select name="{{ form.language.name }}" class="ui dropdown">
+            {% for choice in form.language.field.choices %}
+                <option value="{{ choice.0 }}">{{ choice.1 }}</option>
+            {% endfor %}
+        </select>
+
+        {% if form.language.errors %}
+            <div class="ui red message">{{ form.language.errors }}</div>
+        {% endif %}
+    </div>
+    <div class="ui orange message">
+        {% trans 'Caution: Submissions get final when they are successfully compiled!' %}
+    </div>
+    <input type="submit" class="ui green button" value="{% trans 'Submit' %}">
+{% endif %}
