+++ apps/accounts/templates/accounts/panel/battle_history.html
@@ -1,6 +1,8 @@
-{% load static i18n %}
+{% extends 'accounts/panel/panel.html' %}
+{% load static i18n ai_filters tz %}
+{% get_current_language_bidi as LANGUAGE_BIDI %}
 
-<form class="ui form" action="{% url 'accounts:challenge_a_team' participation_id %}" method="post">
+{% block panel_content %}
     <h2 style="margin-top: 20px;">{% trans 'Battle history' %}</h2>
     {% if battle_history|length %}
         <table class="ui center aligned table table-bordered">
@@ -17,7 +19,11 @@
             <tbody>
             {% for match in battle_history %}
                 <tr>
-                    <td>{{ match.time }}</td>
+                    {% if LANGUAGE_BIDI %}
+                    <td>{{ match.time.date|jalali }} - {{ match.time|time }}</td>
+                    {% else %}
+                    <td>{{ match.time.date }} - {{ match.time|time }}</td>
+                    {% endif %}
                     {% if match.part1.depend == participation %}
                         <td>{{ match.part2.depend.team.name }}</td>
                         <td>{% if match.single_matches.first.log %}{{ match.score1 }}{% endif %}</td>
@@ -56,34 +62,7 @@
                 <tr>
                     <th colspan="6">
                         <div class="ui right floated pagination menu">
-                            {% if battle_history.has_previous %}
-                                <a class="icon item"
-                                   href="?page={{ page }}&battles_page={{ battle_history.previous_page_number }}">
-                                    <i class="{% if LANGUAGE_BIDI %}right{% else %}left{% endif %} chevron icon"></i>
-                                </a>
-                            {% else %}
-                                <a class="disabled icon item">
-                                    <i class="{% if LANGUAGE_BIDI %}right{% else %}left{% endif %} chevron icon"></i>
-                                </a>
-                            {% endif %}
-                            {% for i in battle_history.paginator.page_range %}
-                                {% if battle_history.number == i %}
-                                    <a class="active item"><span>{{ i }}</span></a>
-                                {% else %}
-                                    <a class="item"
-                                       href="?page={{ page }}&battles_page={{ i }}"><span>{{ i }}</span></a>
-                                {% endif %}
-                            {% endfor %}
-                            {% if battle_history.has_next %}
-                                <a class="icon item"
-                                   href="?page={{ page }}&battles_page={{ battle_history.next_page_number }}">
-                                    <i class="{% if LANGUAGE_BIDI %}left{% else %}right{% endif %} chevron icon"></i>
-                                </a>
-                            {% else %}
-                                <a class="disabled icon item">
-                                    <i class="{% if LANGUAGE_BIDI %}left{% else %}right{% endif %} chevron icon"></i>
-                                </a>
-                            {% endif %}
+                            {% include 'accounts/panel/battle_history_paginator.html' %}
                         </div>
                     </th>
                 </tr>
@@ -91,31 +70,34 @@
             {% endif %}
         </table>
     {% endif %}
-    {% csrf_token %}
-    <div class="two fields">
-        <div class="field">
-            {{ form_challenge.battle_team.errors }}
-            <label for="{{ form_challenge.battle_team.id_for_label }}">{% trans 'Versus:' %}
-            </label>
-            <select id="{{ form_challenge.battle_team.id_for_label }}" name="{{ form_challenge.battle_team.name }}"
-                    class="ui search dropdown">
-                {% for choice in form_challenge.battle_team.field.choices %}
-                    <option value="{{ choice.0 }}">{{ choice.1 }}</option>
-                {% endfor %}
-            </select>
+    <form class="ui form" action="{% url 'accounts:challenge_a_team' participation_id %}" method="post"
+          style="width: 100% !important;">
+        {% csrf_token %}
+        <div class="two fields">
+            <div class="field">
+                {{ form_challenge.battle_team.errors }}
+                <label for="{{ form_challenge.battle_team.id_for_label }}">{% trans 'Versus:' %}
+                </label>
+                <select id="{{ form_challenge.battle_team.id_for_label }}" name="{{ form_challenge.battle_team.name }}"
+                        class="ui search dropdown">
+                    {% for choice in form_challenge.battle_team.field.choices %}
+                        <option value="{{ choice.0 }}">{{ choice.1 }}</option>
+                    {% endfor %}
+                </select>
+            </div>
+            <div class="field">
+                {{ form_challenge.battle_team_maps.errors }}
+                <label for="{{ form_challenge.battle_team_maps.id_for_label }}">{% trans 'Map:' %}
+                </label>
+                <select id="{{ form_challenge.battle_team_maps.id_for_label }}"
+                        name="{{ form_challenge.battle_team_maps.name }}" class="ui dropdown">
+                    {% for choice in form_challenge.battle_team_maps.field.choices %}
+                        <option value="{{ choice.0 }}">{{ choice.1 }}</option>
+                    {% endfor %}
+                </select>
+            </div>
         </div>
-        <div class="field">
-            {{ form_challenge.battle_team_maps.errors }}
-            <label for="{{ form_challenge.battle_team_maps.id_for_label }}">{% trans 'Map:' %}
-            </label>
-            <select id="{{ form_challenge.battle_team_maps.id_for_label }}"
-                    name="{{ form_challenge.battle_team_maps.name }}" class="ui dropdown">
-                {% for choice in form_challenge.battle_team_maps.field.choices %}
-                    <option value="{{ choice.0 }}">{{ choice.1 }}</option>
-                {% endfor %}
-            </select>
-        </div>
-    </div>
-    <input type="submit" class="ui green button" value="{% trans 'Submit' %}">
-    <a class="ui red button" href="{% url 'game:scoreboard' friendly_competition.id %}"> {% trans 'Scoreboard' %} </a>
-</form>
+        <input type="submit" class="ui green button" value="{% trans 'Submit' %}">
+    </form>
+
+{% endblock %}
