+++ apps/accounts/templates/accounts/panel/team_management.html
@@ -1,62 +1,69 @@
-{% load i18n %}
-{% load static %}
+{% extends 'accounts/panel/panel.html' %}
+{% load static i18n %}
 {% get_current_language_bidi as LANGUAGE_BIDI %}
 
-<h2>{% trans 'Team Status' %}</h2>
-<div>
-    <table class="ui center aligned table">
-        <thead>
-        <tr>
-            <th>{% trans 'Username' %}</th>
-            <th></th>
-        </tr>
-        </thead>
-        <tbody>
-        {% for user, removable in participation_members %}
+{% block panel_content %}
+    <h2>{% trans 'Team Status' %}</h2>
+    <div>
+    {% if accepted_participations|length > 0 %}
+        <table class="ui center aligned table">
+            <thead>
             <tr>
-                <td>{{ user }}</td>
-                <td>
-                    {% if removable %}
+                <th>{% trans 'Username' %}</th>
+                <th></th>
+            </tr>
+            </thead>
+            <tbody>
+            {% for user, removable in participation_members %}
+                <tr>
+                    <td>{{ user }}</td>
+                    <td>
+                        {% if removable %}
 
-                        <a href="{% url 'accounts:cancel_participation_request' participation.id user.id %}"
-                           class="ui right aligned red button tiny">
-                            <i class="remove icon"></i>
-                            {% trans 'Remove' %}
-                        </a>
-                    {% else %}
-                        <span class="ui right aligned green button tiny">
+                            <a href="{% url 'accounts:cancel_participation_request' participation.id user.id %}"
+                               class="ui right aligned red button tiny">
+                                <i class="remove icon"></i>
+                                {% trans 'Remove' %}
+                            </a>
+                        {% else %}
+                            <span class="ui right aligned green button tiny">
                                     <i class="checkmark icon"></i>
                                     {% trans 'Accepted' %}
                                 </span>
-                    {% endif %}
-                </td>
-            </tr>
-        {% endfor %}
-        {% if participation_members|length < participation.challenge.team_size %}
+                        {% endif %}
+                    </td>
+                </tr>
+            {% endfor %}
+            {% if participation_members|length < participation.challenge.team_size %}
+                <tr>
+                    <td>
+                        <a href="{% url 'accounts:add_participation' participation.id %}"
+                           class="ui right aligned green button tiny">
+                            <i class="add icon"></i>
+                            {% trans 'Add a member' %}
+                        </a>
+                    </td>
+                    <td></td>
+                </tr>
+            {% endif %}
+            </tbody>
+            <tfoot>
             <tr>
-                <td>
-                    <a href="{% url 'accounts:add_participation' participation.id %}"
-                       class="ui right aligned green button tiny">
-                        <i class="add icon"></i>
-                        {% trans 'Add a member' %}
-                    </a>
-                </td>
-                <td></td>
-            </tr>
-        {% endif %}
-        </tbody>
-        <tfoot>
-        <tr>
-            <th colspan="2">
-                {% if participation %}
-                    {% if participation.all_members_accepted %}
-                        {% trans "All users have accepted the challenge." %}
-                    {% else %}
-                        {% trans "All users haven't accepted the challenge yet!" %}
+                <th colspan="2">
+                    {% if participation %}
+                        {% if participation.all_members_accepted %}
+                            {% trans "All users have accepted the challenge." %}
+                        {% else %}
+                            {% trans "All users haven't accepted the challenge yet!" %}
+                        {% endif %}
                     {% endif %}
-                {% endif %}
-            </th>
-        </tr>
-        </tfoot>
-    </table>
-</div>
+                </th>
+            </tr>
+            </tfoot>
+        </table>
+
+    {% else %}
+        {% trans 'You have not participated in this challenge yet.' %}
+    {% endif %}
+    </div>
+{% endblock %}
