+++ apps/accounts/templates/accounts/teams/success_create_team.html
