+++ apps/accounts/urls.py
@@ -11,7 +11,7 @@ urlpatterns = [
     url(r'^activate/(?P<uidb64>[0-9A-Za-z_\-]+)/(?P<token>[0-9A-Za-z]{1,13}-[0-9A-Za-z]{1,20})/$',
         views.activate, name='activate'),
     url(r'^create_team/success_create_team$', views.success_create_team, name='success_create_team'),
-    url(r'^panel/(?P<participation_id>[0-9]+)?$', views.panel, name='panel'),
+    url(r'^panel/(?P<participation_id>[0-9]+)?$', views.team_management, name='panel'),
     url(r'^panel/accept_participation/(?P<participation_id>[0-9]+)$', views.accept_participation,
         name='accept_participation'),
     url(r'^panel/reject_participation/(?P<participation_id>[0-9]+)$', views.reject_participation,
@@ -27,5 +27,10 @@ urlpatterns = [
     url(r'^email_sent$', views.email_sent, name='email_sent'),
     url(r'^email_confirm$', views.email_confirm, name='email_confirm'),
     url(r'^email_invalid$', views.email_invalid, name='email_invalid'),
-    url(r'^challenge_a_team/(?P<participation_id>\d+)$', views.challenge_a_team, name='challenge_a_team')
+    url(r'^challenge_a_team/(?P<participation_id>\d+)$', views.challenge_a_team, name='challenge_a_team'),
+
+    url(r'^panel/submissions', views.submissions, name='panel_submissions'),
+    url(r'^panel/team$', views.team_management, name='panel_team_management'),
+    url(r'^panel/bhistory', views.battle_history, name='panel_battle_history'),
+    url(r'^panel/teampc/(?P<team_pc>\d+)$', views.change_team_pc, name='panel_change_team_pc')
 ]
