+++ apps/accounts/views/__init__.py
@@ -1,0 +1,4 @@
+from .authentication import *
+from .teams import *
+from .panel import *
+from .profile import *
