+++ apps/game/models/__init__.py
@@ -1,3 +1,2 @@
 from .competition import *
 from .game import *
-# from .challenge import Challenge, TeamParticipatesChallenge, UserAcceptsTeamInChallenge, TeamSubmission
